refactor: rename ball constants file
